+++ profiles_api/serializers.py
@@ -41,15 +41,6 @@ class ProfileFeedItemSerializer(serializers.ModelSerializer):
         extra_kwargs = {'user_profile': {'read_only': True}}
 
 
-class SubTopicSerializer(serializers.ModelSerializer):
-    """Serializes subtopics"""
-
-    class Meta:
-        model = models.Subtopic
-        fields = ('id', 'user_profile', 'name', 'html', 'topic')
-        extra_kwargs = {'user_profile': {'read_only': True}}
-
-
 class QuestionSerializer(serializers.ModelSerializer):
     """Serializes questions"""
