+++ profiles_api/urls.py
@@ -5,6 +5,7 @@ from rest_framework.routers import DefaultRouter
 from profiles_api import views
 
 from profiles_api.topic import topic_api_view
+from profiles_api.subtopic import subtopic_api_view
 
 
 router = DefaultRouter()
@@ -12,7 +13,7 @@ router.register('hello-viewset', views.HelloViewSet, base_name='hello-viewset')
 router.register('profile', views.UserProfileViewSet)
 router.register('feed', views.UserProfileFeedViewSet)
 router.register('question', views.QuestionViewSet)
-router.register('subtopic', views.SubtopicViewSet)
+router.register('subtopic', subtopic_api_view.SubtopicViewSet)
 router.register('topic', topic_api_view.TopicViewSet)
 router.register('answer', views.AnswerViewSet)
 router.register('test', views.TestViewSet)
@@ -23,7 +24,7 @@ router.register('theoryPage', views.TheoryPageViewSet)
 urlpatterns = [
     path('hello-view/', views.HelloApiView.as_view()),
     path('login/', views.UserLoginApiView.as_view()),
-    path('custom-subtopic/', views.CustomSubtopicView.as_view()),
+    path('custom-subtopic/', subtopic_api_view.CustomSubtopicView.as_view()),
     path('custom-question/', views.QuestionView.as_view()),
     path('custom-topic/', topic_api_view.TopicView.as_view()),
     path('custom-answer/', views.AnswerView.as_view()),
