+++ profiles_api/views.py
@@ -8,14 +8,12 @@ from rest_framework.authtoken.views import ObtainAuthToken
 from rest_framework.settings import api_settings
 from rest_framework.permissions import IsAuthenticated
 
-from rest_framework.authtoken.models import Token #Get user_profile from corresponding token
-
 from profiles_api import serializers
 from profiles_api import models
 from profiles_api import permissions
 
 from profiles_api.topic.topic_model import Topic
-
+from profiles_api.subtopic.subtopic_model import Subtopic
 
 
 class HelloApiView(APIView):
@@ -168,43 +166,6 @@ class QuestionViewSet(viewsets.ModelViewSet):
         return questions
 
 
-class SubtopicViewSet(viewsets.ModelViewSet):
-    """Handles creating, reading and updating subtopics"""
-    authentication_classes = (TokenAuthentication,)
-    serializer_class = serializers.SubTopicSerializer
-    queryset = models.Subtopic.objects.all()
-    permission_classes = (permissions.UpdateOwnStatus, IsAuthenticated)
-
-    def perform_create(self, serializer):
-        """Sets the user profile to the logged in user"""
-        serializer.save(user_profile=self.request.user)
-
-    def get_queryset(self):
-        """Retrieve only subtopic with certain topic"""
-        topic = self.request.query_params.get('topic', None)
-        topic_id = self.request.query_params.get('topic_id', None)
-        start = self.request.query_params.get('start', None)
-        number = self.request.query_params.get('number', None)
-
-        """Return subtopics"""
-        if topic is not None:
-            filter_dict = {'topic__name': topic}
-            subtopics = models.Subtopic.objects.filter(**filter_dict)
-        elif topic_id is not None:
-            filter_dict = {'topic__id': topic_id}
-            subtopics = models.Subtopic.objects.filter(**filter_dict)
-        else:
-            subtopics = models.Subtopic.objects.all()
-
-        if start is not None:
-            subtopics = subtopics[min(abs(int(start)), subtopics.count()):]
-
-        if number is not None:
-            subtopics = subtopics[:max(0, min(int(number), subtopics.count()))]
-
-        return subtopics
-
-
 class AnswerViewSet(viewsets.ModelViewSet):
     """Handles creating, reading and updating answers"""
     authentication_classes = (TokenAuthentication,)
@@ -326,9 +287,9 @@ class QuestionView(APIView):
                 return Response(data='Subtopic not defined', status=400)
             else:
                 filter_dict = {'id': subtopic_id}
-                subtopic = models.Subtopic.objects.filter(**filter_dict)[0]
+                subtopic = Subtopic.objects.filter(**filter_dict)[0]
         else:
-            subtopic = models.Subtopic.objects.get_or_create(
+            subtopic = Subtopic.objects.get_or_create(
                 name=subtopic_name,
                 topic=topic,
                 user_profile_id=user.id
@@ -355,7 +316,7 @@ class QuestionView(APIView):
             for dependency_string in dependencies_string:
                 filter_dict = {'name': dependency_string}
                 try:
-                    dependency = models.Subtopic.objects.filter(**filter_dict)[0]
+                    dependency = Subtopic.objects.filter(**filter_dict)[0]
                     question.dependencies.add(dependency)
                 except:
                     pass
@@ -365,7 +326,7 @@ class QuestionView(APIView):
             for dependency_string in dependencies_string:
                 filter_dict = {'id': dependency_string}
                 try:
-                    dependency = models.Subtopic.objects.filter(**filter_dict)[0]
+                    dependency = Subtopic.objects.filter(**filter_dict)[0]
                     question.dependencies.add(dependency)
                 except:
                     pass
@@ -380,67 +341,6 @@ class QuestionView(APIView):
             question.imageSrc = imageSrc
 
         serializer = serializers.QuestionSerializer(question)
-        return Response(data=serializer.data, status=201)
-
-
-class CustomSubtopicView(APIView):
-    """Custom view for Subtopic"""
-    authentication_classes = (TokenAuthentication,)
-    permission_classes = (permissions.UpdateOwnStatus, IsAuthenticated)
-
-    def get(self, request):
-        """Retrieve only certain subtopics"""
-        topic = self.request.query_params.get('topic', None)
-        topic_id = self.request.query_params.get('topic_id', None)
-        start = self.request.query_params.get('start', None)
-        number = self.request.query_params.get('number', None)
-
-        if topic is not None:
-            filter_dict = {'topic__name': topic}
-            subtopics = models.Subtopic.objects.filter(**filter_dict)
-        elif topic_id is not None:
-            filter_dict = {'topic__id': topic_id}
-            subtopics = models.Subtopic.objects.filter(**filter_dict)
-        else:
-            subtopics = models.Subtopic.objects.all()
-
-        if start is not None:
-            subtopics = subtopics[min(abs(int(start)), subtopics.count()):]
-
-        if number is not None:
-            subtopics = subtopics[:max(0, min(int(number), subtopics.count()))]
-
-        if subtopics.count() > 0:
-            serializer = serializers.SubTopicSerializer(subtopics, many=True)
-            return Response(data=serializer.data, status=200)
-        else:
-            return Response(status=204)
-
-    def post(self, request):
-        """Create a new subtopic. Additionally a new topic is created if necessary"""
-        topic_name = request.data['topic']
-        if topic_name is None or topic_name == '':
-            return Response(data='Topic not defined', status=400)
-
-        user = self.request.user
-
-        if user is None or user.id == '':
-            return Response(data='User not defined', status=400)
-
-        topic = Topic.objects.get_or_create(
-            name=topic_name,
-            user_profile_id=user.id
-        )[0]
-
-        subtopic = models.Subtopic.objects.get_or_create(
-            topic=topic,
-            name=request.data['name'],
-            html=request.data['html'],
-            user_profile_id=user.id
-        )[0]
-
-        serializer = serializers.SubTopicSerializer(subtopic)
-
         return Response(data=serializer.data, status=201)
